Implement the Issue class
This class is used to represent an JIRA issue.
It has a `key` (similar to a GitHubObject's `name`), and a `status` attributes.
It also contains methods to retrieve JIRA issues through the JIRA API,
a method to generate a JIRA issue key from an object (using the object's name
and a Transform object),
and to create Issue objects from the json API responses.