Extract logic of creating thread pools and determining their size.
Previously, the logic of creating thread pools, and determining the
number of threads in each pool was done explicitly inside the functions
that interact with the github and jira APIs.

Now the logic behind these operation is performed inside designated
separate functions, that are called within the functions interacting
with the APIs.

This makes the code more modular, and makes for easier and better
testing.