Add __eq__ and __ne__ to model.Issue
Previously, two Issues were compared by their memory addresses.
This was not intentional, just a consequence of not implementing
__eq__ and __ne__ because there was no need to do so.

Now, as part of writing tests for the Issue class, a need for
the implementation has emerged. The comparison is done by comparing
the dicts of the Issue, just as in GitHubObject and its subclasses.