Change model.Precedence.__set__ exception-raising
Previously, model.Precedence.__set__ raised a TypeError if the
value assigned to that Precedence was not of type `int`, or was not
positive.

Raising a TypeError when given a float while expecting an int seems
like correct behavior. However, Raising a TypeError when given a
non-positive int while expecting a positive int is not a type issue,
but more of a value issue.

With these considerations in mind, I decided to spilt the exception-
raising clause into two, and explicitly raise a ValueError when given
a positive integer.
(getting an non-integer object still raises a TypeError)